chore: Add IQOptionAccount
