+++ examples/iqoption/getProfile.js
@@ -1,10 +1,10 @@
 import '../../loadEnv'
 
 import { Hemes } from '@hemes/core'
-import { IQOptionProvider, BaseIQOptionProvider } from '@hemes/iqoption'
+import { IQOptionProvider } from '@hemes/iqoption'
 
 async function run() {
-  const hemes = new Hemes(IQOptionProvider).getProvider<BaseIQOptionProvider>()
+  const hemes = new Hemes(IQOptionProvider).getProvider()
 
   const account = await hemes.logIn({
     email: String(process.env.TEST_IQOPTION_ACCOUNT_EMAIL),
